refactor: move hook to a package
