+++ cmd/internal/hook/testdata/sleeping_beauty.sh
