+++ cmd/internal/hook/testdata/sleepy.sh
